build: modernize the build system
This makes the smallest chanages possible to the source code to make it use ES modules. It does not attempt to use them properly, but instead keeps the object namespace approach the library used previously.

It switches the build system from an Ant build config to using Rollup, and switches from yui-compressor to terser for minification.